Implement OpenMP for macos clang
This is a preliminary implementation specifically for mac (hence all the imports in all the pro files that require sala) because apple's clang doesn't really have openmp exposed yet. For this to work, llvm has to be installed through homebrew, and a relevant kit created in qt creator with clang from that llvm as compiler.